==> 02-randomize/00-join_ties_randomization.ipynb <==
import pandas as pd
from tweetple import TweetPle
import sys
sys.path.insert(0, '../../src/utils')
from funcs import *
from tqdm import tqdm
sys.path.insert(0, '../../src/utils')
from general import *

base = f'../../data/03-experiment/SA/baseline/'
path_tw = base + '00-raw/influencers/tweets/'
path_likes = f'{base}00-raw/influencers/likes/'

tweets = pd.read_parquet('../../data/03-experiment/KE/baseline/01-preprocess/influencers/tweets_batch2.parquet')

followers_sa = pd.read_parquet('../../data/03-experiment/SA/baseline/01-preprocess/influencers/followers_ties_batch2.parquet')

followers_grouped_n_sa = followers_sa[['influencer_id', 'follower_id']].groupby('influencer_id').count()
followers_grouped_n_sa.rename({'follower_id': 'followers'}, axis=1, inplace=True)

followers_grouped_sa = followers_sa.groupby('influencer_id').sum()
followers_grouped_sa = followers_grouped_sa.merge(followers_grouped_n_sa, on='influencer_id', how='left')
followers_grouped_sa = followers_grouped_sa[['followers','strong', 'weak']]
followers_grouped_sa['absent'] = followers_grouped_sa['followers'] - followers_grouped_sa['strong'] - followers_grouped_sa['weak']

followers_ke = pd.read_parquet('../../data/03-experiment/KE/baseline/01-preprocess/influencers/followers_ties_batch2.parquet')

followers_grouped_n = followers_ke[['influencer_id', 'follower_id']].groupby('influencer_id').count()
followers_grouped_n.rename({'follower_id': 'followers'}, axis=1, inplace=True)

followers_grouped_ke = followers_ke.groupby('influencer_id').sum()
followers_grouped_ke = followers_grouped_ke.merge(followers_grouped_n, on='influencer_id', how='left')

followers_grouped_ke = followers_grouped_ke[['followers','strong', 'weak']]
followers_grouped_ke['absent'] = followers_grouped_ke['followers'] - followers_grouped_ke['strong'] - followers_grouped_ke['weak']

rand_vars_ke = pd.read_parquet('../../data/02-randomize/KE/02-variables/variables_batch2.parquet')
rand_vars_sa = pd.read_parquet('../../data/02-randomize/SA/02-variables/variables_batch2.parquet')

followers_grouped_ke.rename(columns = {'followers':'n_followers',
                            'strong':'n_strong',
                            'weak':'n_weak',
                            'absent':'n_absent'}, inplace = True)
followers_grouped_sa.rename(columns = {'followers':'n_followers',
                            'strong':'n_strong',
                            'weak':'n_weak',
                            'absent':'n_absent'}, inplace = True)

followers_grouped_sa.reset_index(inplace=True)
followers_grouped_ke.reset_index(inplace=True)

followers_grouped_ke.rename(columns = {'influencer_id':'author_id'}, inplace = True)
followers_grouped_sa.rename(columns = {'influencer_id':'author_id'}, inplace = True)

followers_grouped_ke['author_id'] = followers_grouped_ke['author_id'].astype(str)
followers_grouped_sa['author_id'] = followers_grouped_sa['author_id'].astype(str)
rand_vars_ke['author_id'] = rand_vars_ke['author_id'].astype(str)
rand_vars_sa['author_id'] = rand_vars_sa['author_id'].astype(str)

rand_vars_ke = pd.merge(rand_vars_ke,followers_grouped_ke,on='author_id',how='left')
rand_vars_sa = pd.merge(rand_vars_sa,followers_grouped_sa,on='author_id',how='left')

rand_vars_sa.columns

# Remove timezone from columns if needed
rand_vars_sa['created_at'] = rand_vars_sa['created_at'].dt.tz_localize(None)
rand_vars_ke['created_at'] = rand_vars_ke['created_at'].dt.tz_localize(None)

rand_vars_sa = rand_vars_sa.drop('Unnamed: 0', axis=1)

rand_vars_ke.to_excel('../../data/02-randomize/KE/02-variables/variables_batch2.xlsx')
rand_vars_sa.to_excel('../../data/02-randomize/SA/02-variables/variables_batch2.xlsx')

##### In case of days_old needed

from datetime import date
today = date.today()
rand_vars_ke['date_today'] = today.strftime("%Y-%m-%d")
rand_vars_sa['date_today'] = today.strftime("%Y-%m-%d")

rand_vars_ke['created_at'] = rand_vars_ke['created_at'].apply(pd.to_datetime) #if conversion required
rand_vars_ke['date_today'] = rand_vars_ke['date_today'].apply(pd.to_datetime) #if conversion required
rand_vars_ke['days_old_account'] = (rand_vars_ke['date_today'] - rand_vars_ke['created_at']).dt.days

rand_vars_sa['created_at'] = rand_vars_sa['created_at'].apply(pd.to_datetime) #if conversion required
rand_vars_sa['date_today'] = rand_vars_sa['date_today'].apply(pd.to_datetime) #if conversion required
rand_vars_sa['days_old_account'] = (rand_vars_sa['date_today'] - rand_vars_sa['created_at']).dt.days

rand_vars_ke.to_excel('../../data/02-randomize/KE/02-variables/variables.xlsx')
rand_vars_sa.to_excel('../../data/02-randomize/SA/02-variables/variables.xlsx')

### Followers Randomization

followers_sa = pd.read_parquet('../../data/03-experiment/SA/baseline/01-preprocess/influencers/followers_ties_batch2.parquet')

# Fixing ties: if strong then weak = 0
followers_sa['weak'] = np.where(followers_sa['strong']== 1, 0, followers_sa['weak'])

followers_grouped_n_sa = followers_sa[['influencer_id', 'follower_id']].groupby('follower_id').count()
followers_grouped_n_sa.rename({'influencer_id': 'n_following'}, axis=1, inplace=True)

followers_grouped_sa = followers_sa.groupby('follower_id').sum()
followers_grouped_sa = followers_grouped_sa.merge(followers_grouped_n_sa, on='follower_id', how='left')
followers_grouped_sa = followers_grouped_sa[['n_following','strong', 'weak']]
followers_grouped_sa['absent'] = followers_grouped_sa['n_following'] - followers_grouped_sa['strong'] - followers_grouped_sa['weak']

followers_ke = pd.read_parquet('../../data/03-experiment/KE/baseline/01-preprocess/influencers/followers_ties_batch2.parquet')

followers_ke['weak'] = np.where(followers_ke['strong']== 1, 0, followers_ke['weak'])

followers_grouped_n_ke = followers_ke[['influencer_id', 'follower_id']].groupby('follower_id').count()
followers_grouped_n_ke.rename({'influencer_id': 'n_following'}, axis=1, inplace=True)

followers_grouped_ke = followers_ke.groupby('follower_id').sum()
followers_grouped_ke = followers_grouped_ke.merge(followers_grouped_n_ke, on='follower_id', how='left')
followers_grouped_ke = followers_grouped_ke[['n_following','strong', 'weak']]
followers_grouped_ke['absent'] = followers_grouped_ke['n_following'] - followers_grouped_ke['strong'] - followers_grouped_ke['weak']

followers_grouped_sa.reset_index(inplace=True)
followers_grouped_ke.reset_index(inplace=True)
followers_grouped_ke.rename(columns = {'followers':'n_followers',
                            'strong':'n_strong',
                            'weak':'n_weak',
                            'absent':'n_absent'}, inplace = True)
followers_grouped_sa.rename(columns = {'followers':'n_followers',
                            'strong':'n_strong',
                            'weak':'n_weak',
                            'absent':'n_absent'}, inplace = True)

followers_grouped_ke['follower_id'] = followers_grouped_ke['follower_id'].astype(str)
followers_grouped_sa['follower_id'] = followers_grouped_sa['follower_id'].astype(str)

b_path = f'../../data/01-characterize/followers/KE/00-raw/'
file = 'integrate/followers_batch2.parquet.gzip'
info_followers_ke = pd.read_parquet(f'{b_path}{file}')

b_path = f'../../data/01-characterize/followers/SA/00-raw/'
file = 'integrate/followers_batch2.parquet.gzip'
info_followers_sa = pd.read_parquet(f'{b_path}{file}')

from ast import literal_eval


info_followers_sa = info_followers_sa.join(
      pd.json_normalize(info_followers_sa['public_metrics'])
 ).drop('public_metrics',axis=1)

info_followers_ke = info_followers_ke.join(
      pd.json_normalize(info_followers_ke['public_metrics'])
 ).drop('public_metrics',axis=1)

info_followers_sa.drop_duplicates('id', keep = 'last', inplace=True)
info_followers_ke.drop_duplicates('id', keep = 'last', inplace=True)

info_followers_ke.reset_index(drop=True, inplace=True)
info_followers_sa.reset_index(drop=True, inplace=True)

ids_ke = list(info_followers_ke['id'].astype(str))
ids_sa = list(info_followers_sa['id'].astype(str))
bearer_token = "AAAAAAAAAAAAAAAAAAAAAFpgZAEAAAAAbJS59UWzipi32ixd7LHtXov9olo%3D7gxD8Afshgj4munMXHLU08jzRdTpsAh4RZqq7VBofq1wAvkx1T"

info_followers_ke.rename(columns={'id':'follower_id',
                                  'public_metrics.followers_count':'followers_count',
                                  'public_metrics.following_count': 'following_count',
                                  'public_metrics.tweet_count':'tweet_count',
                                  'public_metrics.listed_count':'listed_count',}, inplace=True)
info_followers_sa.rename(columns={'id':'follower_id',
                                  'public_metrics.followers_count':'followers_count',
                                  'public_metrics.following_count': 'following_count',
                                  'public_metrics.tweet_count':'tweet_count',
                                  'public_metrics.listed_count':'listed_count',}, inplace=True)

info_followers_ke = info_followers_ke.merge(followers_grouped_ke,on='follower_id',how='left')
info_followers_sa = info_followers_sa.merge(followers_grouped_sa,on='follower_id',how='left')

info_followers_ke.to_parquet('../../data/02-randomize/KE/02-variables/variables_followers_batch2.parquet')
info_followers_sa.to_parquet('../../data/02-randomize/SA/02-variables/variables_followers_batch2.parquet')

#### ???????

treatment_ke = pd.read_excel('../../data/02-randomize/KE/03-assignment/output/RandomizedTwitterSampleKE_batch2.xlsx')
treatment_sa = pd.read_excel('../../data/02-randomize/SA/03-assignment/output/RandomizedTwitterSampleSA_batch2.xlsx')

treatment_ke['author_id'] = treatment_ke['author_id'].astype(str)
treatment_sa['author_id'] = treatment_sa['author_id'].astype(str)

control_ke = list(treatment_ke[treatment_ke['treatment']==0].author_id)
control_sa = list(treatment_sa[treatment_sa['treatment']==0].author_id)

treat_ke = list(treatment_ke[treatment_ke['treatment']==1].author_id)
treat_sa = list(treatment_sa[treatment_sa['treatment']==1].author_id)

info_followers_ke = pd.read_parquet('../../data/02-randomize/KE/02-variables/variables_followers_batch2.parquet')
info_followers_sa = pd.read_parquet('../../data/02-randomize/SA/02-variables/variables_followers_batch2.parquet')



followers_sa = pd.read_parquet('../../data/03-experiment/SA/baseline/01-preprocess/influencers/followers_ties_batch2.parquet')
followers_ke =  pd.read_parquet('../../data/03-experiment/KE/baseline/01-preprocess/influencers/followers_ties_batch2.parquet')

followers_ke['weak'] = np.where(followers_ke['strong']== 1, 0, followers_ke['weak'])
followers_sa['weak'] = np.where(followers_sa['strong']== 1, 0, followers_sa['weak'])

followers_sa = followers_sa[followers_sa['influencer_id'].isin(treat_sa)]
followers_ke = followers_ke[followers_ke['influencer_id'].isin(treat_ke)]

followers_grouped_t_sa = followers_sa[['influencer_id', 'follower_id']].groupby('follower_id').count()
followers_grouped_t_sa.rename({'influencer_id': 'n_following_treated'}, axis=1, inplace=True)


followers_grouped_sa_t = followers_sa.groupby('follower_id').sum()
followers_grouped_sa_t = followers_grouped_sa_t.merge(followers_grouped_t_sa, on='follower_id', how='left')
followers_grouped_sa_t = followers_grouped_sa_t[['n_following_treated','strong', 'weak']]
followers_grouped_sa_t['absent'] = followers_grouped_sa_t['n_following_treated'] - followers_grouped_sa_t['strong'] - followers_grouped_sa_t['weak']



followers_grouped_t_ke = followers_ke[followers_ke['influencer_id'].isin(treat_ke)]
followers_grouped_t_ke = followers_grouped_t_ke[['influencer_id', 'follower_id']].groupby('follower_id').count()
followers_grouped_t_ke.rename({'influencer_id': 'n_following_treated'}, axis=1, inplace=True)


followers_grouped_ke_t = followers_ke[followers_ke['influencer_id'].isin(treat_ke)]
followers_grouped_ke_t = followers_grouped_ke_t.groupby('follower_id').sum()
followers_grouped_ke_t = followers_grouped_ke_t.merge(followers_grouped_t_ke, on='follower_id', how='left')
followers_grouped_ke_t = followers_grouped_ke_t[['n_following_treated','strong', 'weak']]
followers_grouped_ke_t['absent'] = followers_grouped_ke_t['n_following_treated'] - followers_grouped_ke_t['strong'] - followers_grouped_ke_t['weak']




followers_grouped_sa_t.reset_index(inplace=True)
followers_grouped_ke_t.reset_index(inplace=True)
followers_grouped_ke_t.rename(columns = {'followers':'n_followers_treated',
                            'strong':'n_strong_treated',
                            'weak':'n_weak_treated',
                            'absent':'n_absent_treated'}, inplace = True)
followers_grouped_sa_t.rename(columns = {'followers':'n_followers_treated',
                            'strong':'n_strong_treated',
                            'weak':'n_weak_treated',
                            'absent':'n_absent_treated'}, inplace = True)

followers_grouped_ke_t['follower_id'] = followers_grouped_ke_t['follower_id'].astype(str)
followers_grouped_sa_t['follower_id'] = followers_grouped_sa_t['follower_id'].astype(str)

info_followers_ke = info_followers_ke.merge(followers_grouped_ke_t,on='follower_id',how='left')
info_followers_sa = info_followers_sa.merge(followers_grouped_sa_t,on='follower_id',how='left')

## control

#info_followers_ke = pd.read_parquet('../../data/02-randomize/KE/02-variables/variables_followers.parquet')
#info_followers_sa = pd.read_parquet('../../data/02-randomize/SA/02-variables/variables_followers.parquet')

followers_sa = pd.read_parquet('../../data/03-experiment/SA/baseline/01-preprocess/influencers/followers_ties_batch2.parquet')
followers_ke =  pd.read_parquet('../../data/03-experiment/KE/baseline/01-preprocess/influencers/followers_ties_batch2.parquet')

followers_ke['weak'] = np.where(followers_ke['strong']== 1, 0, followers_ke['weak'])
followers_sa['weak'] = np.where(followers_sa['strong']== 1, 0, followers_sa['weak'])

followers_sa = followers_sa[followers_sa['influencer_id'].isin(control_sa)]
followers_ke = followers_ke[followers_ke['influencer_id'].isin(control_ke)]

followers_grouped_c_sa = followers_sa[['influencer_id', 'follower_id']].groupby('follower_id').count()
followers_grouped_c_sa.rename({'influencer_id': 'n_following_control'}, axis=1, inplace=True)


followers_grouped_sa_c = followers_sa.groupby('follower_id').sum()
followers_grouped_sa_c = followers_grouped_sa_c.merge(followers_grouped_c_sa, on='follower_id', how='left')
followers_grouped_sa_c = followers_grouped_sa_c[['n_following_control','strong', 'weak']]
followers_grouped_sa_c['absent'] = followers_grouped_sa_c['n_following_control'] - followers_grouped_sa_c['strong'] - followers_grouped_sa_c['weak']

followers_grouped_c_ke = followers_ke[followers_ke['influencer_id'].isin(control_ke)]
followers_grouped_c_ke = followers_grouped_c_ke[['influencer_id', 'follower_id']].groupby('follower_id').count()
followers_grouped_c_ke.rename({'influencer_id': 'n_following_control'}, axis=1, inplace=True)


followers_grouped_ke_c = followers_ke[followers_ke['influencer_id'].isin(control_ke)]
followers_grouped_ke_c = followers_grouped_ke_c.groupby('follower_id').sum()
followers_grouped_ke_c = followers_grouped_ke_c.merge(followers_grouped_c_ke, on='follower_id', how='left')
followers_grouped_ke_c = followers_grouped_ke_c[['n_following_control','strong', 'weak']]
followers_grouped_ke_c['absent'] = followers_grouped_ke_c['n_following_control'] - followers_grouped_ke_c['strong'] - followers_grouped_ke_c['weak']


followers_grouped_sa_c.reset_index(inplace=True)
followers_grouped_ke_c.reset_index(inplace=True)
followers_grouped_ke_c.rename(columns = {'followers':'n_followers_control',
                            'strong':'n_strong_control',
                            'weak':'n_weak_control',
                            'absent':'n_absent_control'}, inplace = True)
followers_grouped_sa_c.rename(columns = {'followers':'n_followers_control',
                            'strong':'n_strong_control',
                            'weak':'n_weak_control',
                            'absent':'n_absent_control'}, inplace = True)

followers_grouped_ke_c['follower_id'] = followers_grouped_ke_c['follower_id'].astype(str)
followers_grouped_sa_c['follower_id'] = followers_grouped_sa_c['follower_id'].astype(str)

info_followers_ke = info_followers_ke.merge(followers_grouped_ke_c,on='follower_id',how='left')
info_followers_sa = info_followers_sa.merge(followers_grouped_sa_c,on='follower_id',how='left')

info_followers_ke.drop('author_id_following', axis=1, inplace=True)
info_followers_sa.drop('author_id_following', axis=1, inplace=True)

info_followers_ke["n_following_treated"] = info_followers_ke["n_following_treated"].fillna(0)
info_followers_ke["n_strong_treated"] = info_followers_ke["n_strong_treated"].fillna(0)
info_followers_ke["n_weak_treated"] = info_followers_ke["n_weak_treated"].fillna(0)
info_followers_ke["n_absent_treated"] = info_followers_ke["n_absent_treated"].fillna(0)
info_followers_ke["n_following_control"] = info_followers_ke["n_following_control"].fillna(0)
info_followers_ke["n_strong_control"] = info_followers_ke["n_strong_control"].fillna(0)
info_followers_ke["n_weak_control"] = info_followers_ke["n_weak_control"].fillna(0)
info_followers_ke["n_absent_control"] = info_followers_ke["n_absent_control"].fillna(0)


info_followers_sa["n_following_treated"] = info_followers_sa["n_following_treated"].fillna(0)
info_followers_sa["n_strong_treated"] = info_followers_sa["n_strong_treated"].fillna(0)
info_followers_sa["n_weak_treated"] = info_followers_sa["n_weak_treated"].fillna(0)
info_followers_sa["n_absent_treated"] = info_followers_sa["n_absent_treated"].fillna(0)
info_followers_sa["n_following_control"] = info_followers_sa["n_following_control"].fillna(0)
info_followers_sa["n_strong_control"] = info_followers_sa["n_strong_control"].fillna(0)
info_followers_sa["n_weak_control"] = info_followers_sa["n_weak_control"].fillna(0)
info_followers_sa["n_absent_control"] = info_followers_sa["n_absent_control"].fillna(0)

info_followers_ke.to_parquet('../../data/02-randomize/KE/02-variables/variables_followers_batch2.parquet')
info_followers_sa.to_parquet('../../data/02-randomize/SA/02-variables/variables_followers_batch2.parquet')

import pytz
from datetime import datetime, timedelta
import pandas as pd
from datetime import date
today = datetime.now(tz=pytz.UTC)
info_followers_ke['date_today'] = today
info_followers_sa['date_today'] = today

info_followers_ke['created_at'] = info_followers_ke['created_at'].apply(pd.to_datetime) #if conversion required
info_followers_ke['date_today'] = info_followers_ke['date_today'].apply(pd.to_datetime) #if conversion required
info_followers_ke['days_old_account'] = (info_followers_ke['date_today'] - info_followers_ke['created_at']).dt.days

info_followers_sa['created_at'] = info_followers_sa['created_at'].apply(pd.to_datetime) #if conversion required
info_followers_sa['date_today'] = info_followers_sa['date_today'].apply(pd.to_datetime) #if conversion required
info_followers_sa['days_old_account'] = (info_followers_sa['date_today'] - info_followers_sa['created_at']).dt.days

ids = list(info_followers_ke[info_followers_ke['followers_count']<0]['follower_id'].astype(str))

pd.set_option('display.max_columns', None)
info_followers_ke[info_followers_ke['followers_count']<0]

import tweetple

from tweetple import TweetPle

# Bearer token accesible via Twitter Developer Academic Research Track
bearer_token='AAAAAAAAAAAAAAAAAAAAAFpgZAEAAAAAbJS59UWzipi32ixd7LHtXov9olo%3D7gxD8Afshgj4munMXHLU08jzRdTpsAh4RZqq7VBofq1wAvkx1T'

# List of handle ids
ids = list(info_followers_ke[info_followers_ke['followers_count']<0].follower_id)


missing_info = TweetPle.TweepleStreamer(ids, bearer_token).user_lookup()

#corregir wrong entries

info_followers_ke.loc[174632,'followers_count'] = 0
info_followers_ke.loc[174632,'following_count'] = 61
info_followers_ke.loc[174632,'listed_count'] = 0
info_followers_ke.loc[174632,'tweet_count'] = 0

info_followers_ke.to_parquet('../../data/02-randomize/KE/02-variables/variables_followers_batch2.parquet')
info_followers_sa.to_parquet('../../data/02-randomize/SA/02-variables/variables_followers_batch2.parquet')

followers_sa = pd.read_parquet('../../data/03-experiment/KE/baseline/01-preprocess/influencers/followers_ties.parquet')

ids = pd.read_parquet(f'../../data/02-randomize/KE/04-stratification/integrate/followers_randomized_batch2.parquet')

info_followers_sa= pd.read_parquet('../../data/02-randomize/KE/02-variables/variables_followers_batch2.parquet')

ids = ids.merge(info_followers_sa, on='follower_id', how='left')

r = ids[['follower_id','username_x', 'username_y', 'n_strong', 'n_weak', 'n_absent']]

len(r[((r['n_weak']==0) & (r['n_strong']==0))])

r[((r['n_weak']==0) & (r['n_strong']==0))]

len(r[~((r['n_weak']==0) & (r['n_strong']==0))])

5754/(22887+5754)
