==> 02-randomize/05-join_strats_rand.ipynb <==
import pandas as pd
import os
import glob
import time
import numpy as np

def get_path(country):
    base = f'../../data/02-randomize/{country}/04-stratification/'
    path_strats = base + 'collect_batch2/'
    path_rand = base + 'integrate/'
    return path_strats, path_rand

path_strats, path_rand = get_path('SA')

all_files = glob.glob(os.path.join(path_strats, "*.parquet"))
df = pd.concat((pd.read_parquet(f) for f in all_files), ignore_index=True)
df

df['treatment'] = pd.to_numeric(df['treatment'])

df.to_parquet(f'{path_rand}followers_randomized_batch2.parquet')
